+++ activerecord/lib/rails/generators/active_record/migration/templates/empty.rb
@@ -1,0 +1,4 @@
+class <%= migration_class_name %> < ActiveRecord::Migration
+  def change
+  end
+end